+++ src/modules/markets/AssetSearchInput.tsx
